# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Laravel Base Project API'
  description: 'Laravel Base Project API Documentation'
  version: 1.0.0
servers:
  -
    url: 'http://my-default-host.com'
    description: 'API Server'
paths:
...
  /api/v1/pomodoro/sessions:
    get:
      tags:
        - 'FocusFlow Pomodoro'
      summary: 'Get pomodoro sessions'
      operationId: 99aed8072ea2a96e1db549cca0506cc0
      parameters:
        -
          name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        -
          name: type
          in: query
          required: false
          schema:
            type: string
            enum:
              - work
              - short-break
              - long-break
      responses:
        200:
          description: 'List of sessions'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PomodoroSession'
      security:
        -
          sanctum: []
    post:
      tags:
        - 'FocusFlow Pomodoro'
      summary: 'Start a new pomodoro session'
      operationId: ebf527fbf34819b32550ebdadeda9c23
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                todo_id:
                  type: integer
                  nullable: true
                task_title:
                  type: string
                  nullable: true
                duration:
                  type: integer
                  example: 25
                type:
                  type: string
                  enum:
                    - work
                    - short-break
                    - long-break
                  example: work
              type: object
      responses:
        201:
          description: 'Session started'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroSession'
      security:
        -
          sanctum: []
  '/api/v1/pomodoro/sessions/{id}/complete':
    post:
      tags:
        - 'FocusFlow Pomodoro'
      summary: 'Complete a pomodoro session'
      operationId: 27379ee5d77b836defd371b5cf830ecf
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Session completed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroSession'
      security:
        -
          sanctum: []
  /api/v1/pomodoro/settings:
    get:
      tags:
        - 'FocusFlow Pomodoro'
      summary: 'Get pomodoro settings'
      operationId: 6fbcd961b3a40f70f4abae1551713ed4
      responses:
        200:
          description: 'User settings'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroSetting'
      security:
        -
          sanctum: []
    put:
      tags:
        - 'FocusFlow Pomodoro'
      summary: 'Update pomodoro settings'
      operationId: ea845661a9ffbcf62449cc434d3a1aab
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                work_duration:
                  type: integer
                short_break_duration:
                  type: integer
                long_break_duration:
                  type: integer
                pomodoros_until_long_break:
                  type: integer
                auto_start_breaks:
                  type: boolean
                auto_start_pomodoros:
                  type: boolean
                sound_enabled:
                  type: boolean
                notification_enabled:
                  type: boolean
                sound_volume:
                  type: number
                  format: float
              type: object
      responses:
        200:
          description: 'Settings updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroSetting'
      security:
        -
          sanctum: []
  /api/v1/pomodoro/today-count:
    get:
      tags:
        - 'FocusFlow Pomodoro'
      summary: "Get today's pomodoro count"
      operationId: 975ac9c4d4429588547754da471b687a
      responses:
        200:
          description: "Today's count"
          content:
            application/json:
              schema:
                properties:
                  count:
                    type: integer
                    example: 5
                type: object
      security:
        -
          sanctum: []
...
components:
  schemas:
...
    PomodoroSession:
      title: 'PomodoroSession Model'
      description: 'FocusFlow Pomodoro Session model'
      required:
        - user_id
        - duration
        - type
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        user_id:
          type: integer
          example: 1
        todo_id:
          type: integer
          example: 1
          nullable: true
        task_title:
          type: string
          example: Development
          nullable: true
        start_time:
          type: string
          format: date-time
          example: '2025-01-21T10:00:00Z'
        end_time:
          type: string
          format: date-time
          nullable: true
        duration:
          type: integer
          example: 25
        type:
          type: string
          enum:
            - work
            - short-break
            - long-break
          example: work
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      type: object
    PomodoroSetting:
      title: 'PomodoroSetting Model'
      description: 'FocusFlow Pomodoro Setting model'
      required:
        - user_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        user_id:
          type: integer
          example: 1
        work_duration:
          type: integer
          example: 25
        short_break_duration:
          type: integer
          example: 5
        long_break_duration:
          type: integer
          example: 15
        pomodoros_until_long_break:
          type: integer
          example: 4
        auto_start_breaks:
          type: boolean
          example: false
        auto_start_pomodoros:
          type: boolean
          example: false
        sound_enabled:
          type: boolean
          example: true
        notification_enabled:
          type: boolean
          example: true
        sound_volume:
          type: number
          format: float
          example: 0.7
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      type: object
...
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Laravel Sanctum token authentication. Enter token in format (Bearer <token>)'
...